Save memoized function calls by instance
Via this, memoized functions become aware of their 'this'-context as
part of their invocations.

This is actually quite dangerous, as you have to flush the memoization
cache the moment the state of the 'this' object changes.

Usually, this would be implemented in a reactive way, but lacking a good
observation mechanism, this is really all we can do to come up with a
generalized memoization mechanism.